oeqa/utils/qemurunner: change the serial runner
[YOCTO #15021]

Create a new runner run_serial_socket whose usage matches the traditional
ssh runner. Its return status is 0 when the command succeeds or a negative
error when it fails. If an error is encountered, it raises an Exception.

The previous serial runner is maintained and marked as deprecated.

(From OE-Core rev: 46911f51c086f48e6f25ed75b96d131760971997)

Signed-off-by: Louis Rannou <lrannou@baylibre.com>
[Luca: small fixes to commit message]
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>